progress

+++ TDM.Core/Frameworks/Supermodel.Tooling/Supermodel.Tooling.SolutionMaker.Cmd/Program.cs
@@ -13,7 +13,7 @@ namespace Supermodel.Tooling.SolutionMaker.Cmd
             //return;
 
             Console.BackgroundColor = ConsoleColor.Black;
-            Console.ForegroundColor = ConsoleColor.Magenta;
+            Console.ForegroundColor = ConsoleColor.Green;
             Console.WriteLine("Supermodel.Core Solution Maker. Version Beta 2.0");
             
             //var solutionMakerParams = SolutionMakerParams.ReadFromConsole();
@@ -28,9 +28,27 @@ namespace Supermodel.Tooling.SolutionMaker.Cmd
                 MobileApi = MobileApiEnum.Native,
                 Database = DatabaseEnum.Sqlite
             };
-            Directory.Delete(Path.Combine(solutionMakerParams.SolutionDirectory, solutionMakerParams.SolutionName), true);
+            
+            var path = solutionMakerParams.CalculateFullPath();
+            if (Directory.Exists(path))
+            {
+                Console.ForegroundColor = ConsoleColor.Magenta;
+                Console.Write("Directory ");
+                Console.ForegroundColor = ConsoleColor.Yellow;
+                Console.Write(path);
+                Console.ForegroundColor = ConsoleColor.Magenta;
+                Console.WriteLine(" already exists.\nWould you like to replace it? (y/n)");
+                
+                var input = Console.ReadLine();
+                if (input == null) return;
+                input = input.Trim().ToLower();
+                if (input != "y") return;
+                Directory.Delete(Path.Combine(solutionMakerParams.SolutionDirectory, solutionMakerParams.SolutionName), true);
+            }
 
             SolutionMaker.CreateSupermodelShell(solutionMakerParams);
+            
+            Console.ForegroundColor = ConsoleColor.Green;
         }
     }
 }
